Fix all test-specific sign-conversion warnings (319 eliminated)
Eliminated all 319 sign-conversion warnings in test code by:
- Casting std::ranges::distance results to size_t
- Casting degree() results to size_t
- Using iter_difference_t casts for std::ranges::next
- Using UL suffix for size_t initializer lists

Modified 17 test files:
- test_compressed_graph_cpo.cpp
- test_dynamic_graph_conversions.cpp
- test_dynamic_graph_cpo_dos.cpp
- test_dynamic_graph_cpo_dous.cpp
- test_dynamic_graph_cpo_voem.cpp
- test_dynamic_graph_cpo_vos.cpp
- test_dynamic_graph_cpo_vous.cpp
- test_dynamic_graph_generic_queries.cpp
- test_dynamic_graph_heterogeneous.cpp
- test_dynamic_graph_mixed_types.cpp
- test_dynamic_graph_mol.cpp
- test_dynamic_graph_stl_algorithms.cpp
- test_dynamic_graph_transformations.cpp
- test_dynamic_graph_type_erasure.cpp
- test_edges_uid_cpo.cpp
- test_find_vertex_cpo.cpp
- test_vertices_cpo.cpp

Result: 0 sign-conversion warnings in test code (down from 319)
Note: 280 sign-conversion warnings remain in graph_cpo.hpp from template
instantiation when tests pass int to functions expecting unsigned types.

+++ include/graph/container/compressed_graph.hpp
@@ -1164,8 +1164,8 @@ public: // Friend functions
     }
     
     // Return range [partition_[pid], partition_[pid+1])
-    const auto begin_vid = static_cast<std::size_t>(g.partition_[pid]);
-    const auto end_vid = static_cast<std::size_t>(g.partition_[pid + 1]);
+    const auto begin_vid = static_cast<std::size_t>(g.partition_[static_cast<std::size_t>(pid)]);
+    const auto end_vid = static_cast<std::size_t>(g.partition_[static_cast<std::size_t>(pid + 1)]);
     
     return vertex_descriptor_view<vertex_iter_type>(begin_vid, end_vid);
   }
