+++ tests/test_dynamic_graph_conversions.cpp
@@ -101,7 +101,7 @@ template<typename G>
 size_t count_edges(const G& g) {
     size_t count = 0;
     for (auto&& u : vertices(g)) {
-        count += std::ranges::distance(edges(g, u));
+        count += static_cast<size_t>(std::ranges::distance(edges(g, u)));
     }
     return count;
 }
@@ -109,7 +109,7 @@ size_t count_edges(const G& g) {
 // Helper: Count vertices
 template<typename G>
 size_t count_vertices(const G& g) {
-    return std::ranges::distance(vertices(g));
+    return static_cast<size_t>(std::ranges::distance(vertices(g)));
 }
 
 // Helper: Check if edge exists
