+++ tests/test_dynamic_graph_cpo_dos.cpp
@@ -1158,7 +1158,7 @@ TEST_CASE("dos CPO integration", "[dynamic_graph][dos][cpo][integration]") {
         
         size_t total_edges = 0;
         for (auto v : vertices(g)) {
-            total_edges += degree(g, v);
+            total_edges += static_cast<size_t>(degree(g, v));
         }
         
         REQUIRE(total_edges == num_edges(g));
